Simplify uhid-tokio crate
