+++ linux/system-daemon/Cargo.toml
@@ -17,7 +17,7 @@ libc = "0.2.62"
 nanoid = "0.2.0"
 serde = "1.0.99"
 serde_derive = "1.0.99"
-slog = "2.5.2"
+slog = { version = "2.5.2", features = ["max_level_trace" ] }
 slog-journald = "2.0.0"
 slog-term = "2.4.1"
 systemd = "0.4.0"
