+++ linux/uhid-tokio/src/uhid_device.rs
@@ -1,25 +1,18 @@
-use std::fs::File;
 use std::io::{self, Write};
-use std::os::unix::io::FromRawFd;
 use std::path::Path;
 
 use futures::{Async, AsyncSink, Poll, Sink, StartSend, Stream};
-use nix;
-use nix::fcntl;
-use nix::libc;
 use slog;
 use slog::Drain;
 use slog_stdlog;
-use tokio::reactor::Handle;
 use tokio_io::AsyncRead;
 
-use character_device::{CharacterDevice, Decoder, Encoder, SyncSink};
-use character_device_file::CharacterDeviceFile;
-use uhid_codec::*;
-use uhid_device_file::UHIDDeviceFile;
+use codec::*;
+use misc_driver::MiscDriver;
+use transport::{Decoder, Encoder, SyncSink, Transport};
 
 pub struct UHIDDevice<T> {
-    inner: CharacterDevice<T, UHIDCodec, UHIDCodec>,
+    inner: Transport<T, Codec, Codec>,
     logger: slog::Logger,
 }
 
@@ -36,42 +29,22 @@ pub struct CreateParams {
     pub data: Vec<u8>,
 }
 
-// ===== impl UHIDDevice =====
-
-impl UHIDDevice<UHIDDeviceFile<CharacterDeviceFile<File>>> {
+impl UHIDDevice<MiscDriver> {
     /// Create a UHID device using '/dev/uhid'
     pub fn create<L: Into<Option<slog::Logger>>>(
-        handle: &Handle,
         params: CreateParams,
         logger: L,
-    ) -> io::Result<UHIDDevice<UHIDDeviceFile<CharacterDeviceFile<File>>>> {
-        Self::create_with_path(Path::new("/dev/uhid"), handle, params, logger)
+    ) -> io::Result<UHIDDevice<MiscDriver>> {
+        Self::create_with_path(Path::new("/dev/uhid"), params, logger)
     }
 
-    /// Create a UHID device, using the specified UHID device file path
+    /// Create a UHID device using the specified character misc-device file path
     pub fn create_with_path<L: Into<Option<slog::Logger>>>(
         path: &Path,
-        handle: &Handle,
         params: CreateParams,
         logger: L,
-    ) -> io::Result<UHIDDevice<UHIDDeviceFile<CharacterDeviceFile<File>>>> {
-        let fd = fcntl::open(
-            path,
-            fcntl::OFlag::from_bits(libc::O_RDWR | libc::O_CLOEXEC | libc::O_NONBLOCK).unwrap(),
-            nix::sys::stat::Mode::from_bits(libc::S_IRUSR | libc::S_IWUSR | libc::S_IRGRP | libc::S_IWGRP).unwrap(),
-        ).map_err(|err| {
-            io::Error::new(
-                io::ErrorKind::Other,
-                format!("Cannot open uhid-cdev {}: {}", path.to_str().unwrap(), err),
-            )
-        })?;
-        let file: File = unsafe { File::from_raw_fd(fd) };
-        let device_file = CharacterDeviceFile::new(file);
-        Ok(Self::create_with(
-            device_file.into_io(handle)?,
-            params,
-            logger,
-        ))
+    ) -> io::Result<UHIDDevice<MiscDriver>> {
+        Ok(Self::create_with(MiscDriver::open(path)?, params, logger))
     }
 }
 
@@ -87,11 +60,12 @@ where
         let logger = logger
             .into()
             .unwrap_or(slog::Logger::root(slog_stdlog::StdLog.fuse(), o!()));
+        let logger = logger.new(o!("uhid_device" => params.name.to_string()));
         let mut device = UHIDDevice {
-            inner: CharacterDevice::new(inner, UHIDCodec, UHIDCodec, logger.clone()),
-            logger,
+            inner: Transport::new(inner, Codec, Codec, logger.clone()),
+            logger: logger.clone(),
         };
-        debug!(device.logger, "Send create device event");
+        debug!(logger, "Sending create device event");
         device
             .inner
             .send(InputEvent::Create {
@@ -106,20 +80,21 @@ where
                 data: params.data,
             })
             .unwrap();
-        debug!(device.logger, "Sent create device event");
+        debug!(logger, "Sent create device event");
         device
     }
 
     /// Send a HID packet to the UHID device
-    pub fn send_input(&mut self, data: &[u8]) -> Result<(), <UHIDCodec as Encoder>::Error> {
-        debug!(self.logger, "Send input event");
+    pub fn send_input(&mut self, data: &[u8]) -> Result<(), <Codec as Encoder>::Error> {
+        debug!(self.logger, "send input");
         self.inner.send(InputEvent::Input {
             data: data.to_vec(),
         })
     }
 
     /// Send a 'destroy' to the UHID device and close it
-    pub fn destroy(mut self) -> Result<(), <UHIDCodec as Encoder>::Error> {
+    pub fn destroy(mut self) -> Result<(), <Codec as Encoder>::Error> {
+        debug!(self.logger, "destroy");
         self.inner.send(InputEvent::Destroy)?;
         self.inner.close()?;
         Ok(())
@@ -127,31 +102,33 @@ where
 }
 
 impl<T: AsyncRead> Stream for UHIDDevice<T> {
-    type Item = <UHIDCodec as Decoder>::Item;
-    type Error = <UHIDCodec as Decoder>::Error;
+    type Item = <Codec as Decoder>::Item;
+    type Error = <Codec as Decoder>::Error;
 
     fn poll(&mut self) -> Poll<Option<Self::Item>, Self::Error> {
+        debug!(self.logger, "Stream::poll");
         self.inner.poll()
     }
 }
 
 impl<T: Write> Sink for UHIDDevice<T> {
-    type SinkItem = <UHIDCodec as Encoder>::Item;
-    type SinkError = <UHIDCodec as Encoder>::Error;
+    type SinkItem = <Codec as Encoder>::Item;
+    type SinkError = <Codec as Encoder>::Error;
 
     fn start_send(&mut self, item: Self::SinkItem) -> StartSend<Self::SinkItem, Self::SinkError> {
-        debug!(self.logger, "start_send");
+        debug!(self.logger, "Sink::start_send");
         self.inner.send(item)?;
         Ok(AsyncSink::Ready)
     }
 
     fn poll_complete(&mut self) -> Poll<(), Self::SinkError> {
-        debug!(self.logger, "poll_complete");
+        debug!(self.logger, "Sink::poll_complete");
         self.inner.flush()?;
         Ok(Async::Ready(()))
     }
 
     fn close(&mut self) -> Result<Async<()>, Self::SinkError> {
+        debug!(self.logger, "Sink::close");
         self.inner.close()?;
         Ok(Async::Ready(()))
     }
